Create Balls and Boxes.py
Problem:
You have NN balls and KK boxes. You want to divide the NN balls into KK boxes such that:

->Each box contains greater≥1 balls.
->No two boxes contain the same number of balls.
Determine if it is possible to do so.



Note:

This is one of  my favourite codechef problem, however mentioned above has time complexity, but it gives exactly what we want. i have work on this problem for quite a long and eventually came up with this code. 
This code give u summation of unique sequence of your desired choice of your combinations.